Modified the represenation of real valued variables such that their "low" and "high" values are -100.0 and 100.0 respectively, and not -PI,PI...
This caused problems for translations, e.g., that were outside the
range -3.14 to 3.14 (but worked fine for torsions and quaternion rotations).
The symptoms included energies like "2.0e05" and populations that were
"converged", causing premature exiting of a docking run.

Ideally, we should set the "low" and "high" based on the docking problem:
we should know the minimum and maximum possible translations, and we know
that torsions are always between -PI and PI.

CVS-ID: rep.h 1.4