(WEH - the previous commit log was interupted ... so here's more detail 	about the use of Coliny)
The Coliny-specific makefile fragments, if uncommented, force blocks
defined by

  #if defined(USING_COLINY)
    ...
  #endif

to be compiled.  There was only one change to AutoDock that was not of this
form: the addition of the eval() method in eval.h/eval.cc that was
needed to generalize the EVAL class.

I have compared this code with my previous incarnation of AutoDock+Coliny,
and I get consistent results up to a point where the objective functions
begin to differ (after many iterations).  I suspect that this is due to
some subtle change in the AutoDock code since my version, or perhaps due
to the use of doubles instead of floats somewhere in the new AutoDock code.

The new command line is

        coliny <solver-type> <num-trials>

AutoDock assumes that a <solver-type>.in file is used to initialize the solvers.
For example, if the user specifies

        coliny ms.ps 1

The multistart local search method with pattern search will be run once,
using the ms.ps.in file as input.  Here's an example of such an input file:

max_neval 1000
ls_frequency 1.0
#output_level verbose
debug 100
#debug_all 1
END-SOLVER
exploratory_move adaptive
#debug 10000

Lines beginning with a '#' are commented out.  The line END-SOLVER denotes that the remaining options are for the local search method (pattern search in this case).

To get more information about Coliny solvers, a user can use the AutoDock command line:

	coliny help

to list all solvers currently supported, and the command line

	coliny <solver-type> help

to list the options for this solver along with their default values.

CVS-ID: Makefile 1.5
CVS-ID: coliny.cc 1.2
CVS-ID: main.cc 1.6