After a first round of compilation using Apple's IDE, Xcode, various warnings have been eliminated (mainly unsigned ints and ints being interchanged); and
After using Apple's Shark tool for profiling source code, the
internal energy calculation has been optimized.

The non-bonded cutoff used in the internal energy calculation has been
reduced from 64 Angstroms to 8 Angstroms.  Most contributions beyond
8 Angstroms are very small, less than -0.001 kcal/mol, even for the
largest atoms. Also, the conversion from double to int used to
be done before the if to decide if we were within the cutoff; now
the square of the distance is used in the comparison, and only if
we are within the cutoff, do we convert from the double to int.

The version checked in here still uses the type array to lookup
the energy of interaction for a nonbond; this level of indirection
can be pre-computed, and this should appear in my next round of checkins

-- Garrett

CVS-ID: Makefile 1.6
CVS-ID: autocomm.h 1.5
CVS-ID: banner.cc 1.3
CVS-ID: call_glss.cc 1.7
CVS-ID: constants.h 1.4
CVS-ID: eintcal.cc 1.3
CVS-ID: eval.cc 1.6
CVS-ID: gs.cc 1.4
CVS-ID: gs.h 1.5
CVS-ID: initautodock.cc 1.3
CVS-ID: intnbtable.cc 1.3
CVS-ID: ls.cc 1.3
CVS-ID: main.cc 1.7
CVS-ID: main.h 1.2
CVS-ID: molstruct.h 1.2
CVS-ID: output_state.cc 1.3
CVS-ID: ranlib.cc 1.3
CVS-ID: readPDBQ.cc 1.4
CVS-ID: rep.cc 1.5
CVS-ID: setflags.cc 1.3
CVS-ID: setflags.h 1.2
CVS-ID: simanneal.cc 1.3
CVS-ID: structs.h 1.3
CVS-ID: support.cc 1.6
CVS-ID: support.h 1.3
CVS-ID: trilinterp.cc 1.3
CVS-ID: typedefs.h 1.2